+++ web/app/services.php
@@ -60,6 +60,7 @@ $app['preferences.repository'] = $app->share(function($app) {
         'show_week_nb' => $app['defaults.show_week_nb'],
         'show_now_indicator' => $app['defaults.show_now_indicator'],
         'list_days' => $app['defaults.list_days'],
+        'default_view' => $app['defaults.default_view'],
     ]);
 
     return $repository;